% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWright.R
\docType{package}
\name{MWright}
\alias{MWright}
\alias{MWright-package}
\title{MWright Package}
\description{
Contains random  number generation, plotting, and estimation algorithms for the two-parameter one-sided and two-sided M-Wright (Mainardi-Wright) family.
The M-Wright distributions naturally generalize widely used one-sided (Airy and half-normal or half-normal)
and symmetric (Airy and Gaussian or normal) models. These are widely studied in time-fractional diffusion processes.
}
\details{
References:

Cahoy and Minkabo (2017)<doi:10.3233/MAS-170388>

Cahoy (2012) <doi:10.1007/s00180-011-0269-x>

Cahoy (2012) <doi:10.1080/03610926.2010.543299>

Cahoy (2011)

Mainardi, Mura, and Pagnini (2010) <doi:10.1155/2010/104505>
}
\author{
Dexter Cahoy \email{cahoyd@uhd.edu}
}
